move functions from Ethereum module
