+++ priv/perf/apps/load_test/lib/child_chain/transaction.ex
@@ -23,7 +23,7 @@ defmodule LoadTest.ChildChain.Transaction do
   alias ExPlasma.Transaction
   alias ExPlasma.Utxo
   alias LoadTest.Connection.ChildChain, as: Connection
-  alias LoadTest.Ethereum.Sync
+  alias LoadTest.Service.Sync
 
   # safe, reasonable amount, equal to the testnet block gas limit
   @lots_of_gas 5_712_388
