+++ priv/perf/apps/load_test/lib/ethereum/ethereum.ex
@@ -20,10 +20,11 @@ defmodule LoadTest.Ethereum do
 
   alias ExPlasma.Encoding
   alias LoadTest.ChildChain.Abi
+  alias LoadTest.Ethereum.Account
   alias LoadTest.Ethereum.NonceTracker
-  alias LoadTest.Ethereum.Sync
   alias LoadTest.Ethereum.Transaction
   alias LoadTest.Ethereum.Transaction.Signature
+  alias LoadTest.Service.Sync
 
   @about_4_blocks_time 120_000
   @poll_timeout 60_000
@@ -121,22 +122,8 @@ defmodule LoadTest.Ethereum do
     Sync.repeat_until_success(f, timeout, "Failed to fetch eth block number")
   end
 
-  @spec fetch_balance(Account.addr_t(), non_neg_integer(), Account.addr_t()) :: non_neg_integer() | :error | nil | map()
-  def fetch_balance(address, amount, currency \\ <<0::160>>) do
-    {:ok, result} =
-      Sync.repeat_until_success(
-        fn ->
-          do_fetch_balance(Encoding.to_hex(address), amount, Encoding.to_hex(currency))
-        end,
-        @poll_timeout,
-        "Failed to fetch childchain balance"
-      )
-
-    result
-  end
-
-  @spec fetch_rootchain_balance(Account.addr_t(), Account.addr_t()) :: non_neg_integer() | no_return()
-  def fetch_rootchain_balance(address, <<0::160>>) do
+  @spec fetch_balance(Account.addr_t(), Account.addr_t()) :: non_neg_integer() | no_return()
+  def fetch_balance(address, <<0::160>>) do
     {:ok, initial_balance} =
       Sync.repeat_until_success(
         fn ->
@@ -152,7 +139,7 @@ defmodule LoadTest.Ethereum do
     initial_balance
   end
 
-  def fetch_rootchain_balance(address, currency) do
+  def fetch_balance(address, currency) do
     Sync.repeat_until_success(
       fn ->
         do_root_chain_get_erc20_balance(address, currency)
@@ -160,127 +147,6 @@ defmodule LoadTest.Ethereum do
       @poll_timeout,
       "Failed to fetch erc20 balance from rootchain"
     )
-  end
-
-  @spec create_transaction(
-          non_neg_integer(),
-          Account.addr_t(),
-          Account.addr_t(),
-          Account.addr_t(),
-          non_neg_integer()
-        ) ::
-          {:ok, [binary()]} | {:error, map()}
-
-  def create_transaction(amount_in_wei, input_address, output_address, currency \\ <<0::160>>, timeout \\ 120_000) do
-    func = fn ->
-      transaction = %WatcherInfoAPI.Model.CreateTransactionsBodySchema{
-        owner: Encoding.to_hex(input_address),
-        payments: [
-          %WatcherInfoAPI.Model.TransactionCreatePayments{
-            amount: amount_in_wei,
-            currency: Encoding.to_hex(currency),
-            owner: Encoding.to_hex(output_address)
-          }
-        ],
-        fee: %WatcherInfoAPI.Model.TransactionCreateFee{currency: Encoding.to_hex(currency)}
-      }
-
-      {:ok, response} =
-        WatcherInfoAPI.Api.Transaction.create_transaction(LoadTest.Connection.WatcherInfo.client(), transaction)
-
-      result = Jason.decode!(response.body)["data"]
-
-      process_transaction_result(result)
-    end
-
-    Sync.repeat_until_success(func, timeout, "Failed to created a transaction")
-  end
-
-  @spec submit_transaction(binary(), binary(), [binary()]) :: map()
-  def submit_transaction(typed_data, sign_hash, private_keys) do
-    signatures =
-      Enum.map(private_keys, fn private_key ->
-        sign_hash
-        |> to_binary()
-        |> signature_digest(private_key)
-        |> Encoding.to_hex()
-      end)
-
-    typed_data_signed = Map.put_new(typed_data, "signatures", signatures)
-
-    Sync.repeat_until_success(
-      fn ->
-        submit_typed(typed_data_signed)
-      end,
-      @poll_timeout,
-      "Failed to submit transaction"
-    )
-  end
-
-  defp to_binary(hex) do
-    hex
-    |> String.replace_prefix("0x", "")
-    |> String.upcase()
-    |> Base.decode16!()
-  end
-
-  defp signature_digest(hash_digest, private_key_binary) do
-    {:ok, <<r::size(256), s::size(256)>>, recovery_id} =
-      :libsecp256k1.ecdsa_sign_compact(
-        hash_digest,
-        private_key_binary,
-        :default,
-        <<>>
-      )
-
-    # EIP-155
-    # See https://github.com/ethereum/EIPs/blob/master/EIPS/eip-155.md
-    base_recovery_id = 27
-    recovery_id = base_recovery_id + recovery_id
-
-    <<r::integer-size(256), s::integer-size(256), recovery_id::integer-size(8)>>
-  end
-
-  defp submit_typed(typed_data_signed) do
-    {:ok, response} = execute_submit_typed(typed_data_signed)
-    decoded_response = Jason.decode!(response.body)["data"]
-
-    case decoded_response do
-      %{"messages" => %{"code" => "submit:utxo_not_found"}} = error ->
-        {:error, error}
-
-      %{"messages" => %{"code" => "operation:service_unavailable"}} = error ->
-        {:error, error}
-
-      %{"txhash" => _} ->
-        {:ok, decoded_response}
-    end
-  end
-
-  defp execute_submit_typed(typed_data_signed) do
-    WatcherInfoAPI.Api.Transaction.submit_typed(LoadTest.Connection.WatcherInfo.client(), typed_data_signed)
-  end
-
-  defp process_transaction_result(result) do
-    case result do
-      %{"code" => "create:client_error"} ->
-        {:error, result}
-
-      %{
-        "result" => "complete",
-        "transactions" => [
-          %{
-            "sign_hash" => sign_hash,
-            "typed_data" => typed_data,
-            "txbytes" => txbytes
-          }
-        ]
-      } ->
-        {:ok, [sign_hash, typed_data, txbytes]}
-
-      error ->
-        {:error, error}
-    end
   end
 
   defp eth_account_get_balance(address) do
@@ -303,37 +169,6 @@ defmodule LoadTest.Ethereum do
       error ->
         error
     end
-  end
-
-  defp do_fetch_balance(address, amount, currency) do
-    response =
-      case account_get_balances(address) do
-        {:ok, response} ->
-          decoded_response = Jason.decode!(response.body)
-          Enum.find(decoded_response["data"], fn data -> data["currency"] == currency end)
-
-        result ->
-          Logger.error("Failed to fetch balance from childchain #{inspect(result)}")
-
-          :error
-      end
-
-    case response do
-      # empty response is considered no account balance!
-      nil when amount == 0 ->
-        {:ok, nil}
-
-      %{"amount" => ^amount} = balance ->
-        {:ok, balance}
-
-      response ->
-        {:error, response}
-    end
-  end
-
-  defp account_get_balances(address) do
-    client = LoadTest.Connection.WatcherInfo.client()
-    WatcherInfoAPI.Api.Account.account_get_balance(client, %{address: address})
   end
 
   defp get_external_data(address, signature, params) do
